@@ -1,8 +1,11 @@
 package dev.xinto.argos.domain.courses
 
 data class DomainCourseGroupSchedule(
+    val id: String,
+    val date: String,
     val day: String,
-    val time: String,
+    val startTime: String,
+    val fullTime: String,
     val room: String,
-    val info: String?
+    val lecturer: String
 )
