@@ -2,5 +2,5 @@ package dev.xinto.argos.domain.messages
 
 data class DomainMessageUser(
     val fullName: String,
-    val uid: String,
+    val uid: String
 )
