@@ -4,11 +4,11 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ApiAttributesUserState(
-    val billingBalance: Int,
-    val libraryBalance: Int,
-    val newsUnread: Int,
-    val messagesUnread: Int,
-    val notificationsUnread: Int,
+    val billingBalance: Int?,
+    val libraryBalance: Int?,
+    val newsUnread: Int?,
+    val messagesUnread: Int?,
+    val notificationsUnread: Int?,
     val surveyExists: Boolean,
     val rechoose: Boolean
 )
